+++ src/main.cc
@@ -12,6 +12,10 @@
 #include "reduct/core/logger.h"
 #include "reduct/storage/storage.h"
 
+#ifdef WITH_CONSOLE
+#include "reduct/console.h"
+#endif
+
 using reduct::api::IApiServer;
 using reduct::asset::IAssetManager;
 using reduct::async::ILoop;
@@ -19,12 +23,9 @@ using reduct::auth::ITokenAuthentication;
 using reduct::core::EnvVariable;
 using reduct::core::Error;
 using reduct::core::Logger;
-using reduct::storage::IStorage;
+using ReductStorage = reduct::storage::IStorage;
 
-#ifdef WITH_CONSOLE
-extern const char _binary_console_zip_start[];
-extern const char _binary_console_zip_end[];
-#endif
+
 
 class Loop : public ILoop {
  public:
@@ -58,13 +59,13 @@ int main() {
   ILoop::set_loop(&loop);
 
 #if WITH_CONSOLE
-  auto console = IAssetManager::BuildFromZip(std::string(_binary_console_zip_start, _binary_console_zip_end));
+  auto console = IAssetManager::BuildFromZip(reduct::kZippedConsole);
 #else
   auto console = IAssetManager::BuildEmpty();
 #endif
 
   IApiServer::Components components{
-      .storage = IStorage::Build({.data_path = data_path}),
+      .storage = ReductStorage::Build({.data_path = data_path}),
       .auth = ITokenAuthentication::Build(api_token),
       .console = std::move(console),
   };