+++ src/reduct/core/logger.h
@@ -1,4 +1,4 @@
-// Copyright 2021 Alexey Timin
+// Copyright 2021-2022 Alexey Timin
 #ifndef REDUCT_CORE_LOGGER_H_
 #define REDUCT_CORE_LOGGER_H_
 
@@ -22,8 +22,7 @@ enum class LogLevels { kNone = 0, kError, kWarning, kInfo, kDebug, kTrace };
 class Logger {
  public:
   template <typename... T>
-  static void Log(LogLevels level, std::string_view message, int line, std::string_view file, std::string_view func,
-                  T &&...args) {
+  static void Log(LogLevels level, std::string_view message, int line, std::string_view file, T &&...args) {
     using Clock = std::chrono::system_clock;
 
     static const std::map<LogLevels, std::pair<std::string, fmt::color> > kLoglevelMap = {
@@ -47,9 +46,8 @@ class Logger {
       auto msg = fmt::vformat(message, fmt::make_format_args(args...));
       const auto &[level_str, color] = kLoglevelMap.at(level);
 
-      auto func_name = log_level_ >= LogLevels::kDebug ? fmt::format(" --> {}", func) : "";
-      std::cout << fmt::format(fmt::fg(color), "{}.{:03d} ({:>5}) {:>7} -- {}:{} {} {}", ss.str(), milliseconds,
-                               reinterpret_cast<uint16_t &>(thid), level_str, file, line, msg, func_name)
+      std::cout << fmt::format(fmt::fg(color), "{}.{:03d} ({:>5}) {:>7} -- {}:{} {}", ss.str(), milliseconds,
+                               reinterpret_cast<uint16_t &>(thid), level_str, file, line, msg)
                 << std::endl;
     }
   }
@@ -70,24 +68,24 @@ constexpr const char *r_slant(const char *str) { return *str == '/' ? (str + 1) 
 
 constexpr const char *file_name(const char *str) { return str_slant(str) ? r_slant(str_end(str)) : str; }
 
-#define LOG_ERROR(msg, ...)                                                                      \
-  reduct::core::Logger::Log(reduct::core::LogLevels::kError, msg, __LINE__, file_name(__FILE__), \
-                            __PRETTY_FUNCTION__ __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
+#define LOG_ERROR(msg, ...)                                                 \
+  reduct::core::Logger::Log(reduct::core::LogLevels::kError, msg, __LINE__, \
+                            file_name(__FILE__) __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
 
-#define LOG_WARNING(msg, ...)                                                                      \
-  reduct::core::Logger::Log(reduct::core::LogLevels::kWarning, msg, __LINE__, file_name(__FILE__), \
-                            __PRETTY_FUNCTION__ __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
+#define LOG_WARNING(msg, ...)                                                 \
+  reduct::core::Logger::Log(reduct::core::LogLevels::kWarning, msg, __LINE__, \
+                            file_name(__FILE__) __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
 
-#define LOG_INFO(msg, ...)                                                                      \
-  reduct::core::Logger::Log(reduct::core::LogLevels::kInfo, msg, __LINE__, file_name(__FILE__), \
-                            __PRETTY_FUNCTION__ __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
+#define LOG_INFO(msg, ...)                                                 \
+  reduct::core::Logger::Log(reduct::core::LogLevels::kInfo, msg, __LINE__, \
+                            file_name(__FILE__) __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
 
-#define LOG_DEBUG(msg, ...)                                                                      \
-  reduct::core::Logger::Log(reduct::core::LogLevels::kDebug, msg, __LINE__, file_name(__FILE__), \
-                            __PRETTY_FUNCTION__ __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
+#define LOG_DEBUG(msg, ...)                                                 \
+  reduct::core::Logger::Log(reduct::core::LogLevels::kDebug, msg, __LINE__, \
+                            file_name(__FILE__) __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
 
-#define LOG_TRACE(msg, ...)                                                                      \
-  reduct::core::Logger::Log(reduct::core::LogLevels::kTrace, msg, __LINE__, file_name(__FILE__), \
-                            __PRETTY_FUNCTION__ __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
+#define LOG_TRACE(msg, ...)                                                 \
+  reduct::core::Logger::Log(reduct::core::LogLevels::kTrace, msg, __LINE__, \
+                            file_name(__FILE__) __VA_OPT__(, ) __VA_ARGS__)  // NOLINT
 
 #endif  //  REDUCT_CORE_LOGGER_H_