+++ src/reduct/storage/storage.cc
@@ -29,7 +29,7 @@ class Storage : public IStorage {
 
     for (const auto& folder : fs::directory_iterator(options_.data_path)) {
       if (folder.is_directory()) {
-        auto bucket_name = folder.path().filename();
+        auto bucket_name = folder.path().filename().string();
         if (auto bucket = IBucket::Restore(folder)) {
           buckets_[bucket_name] = std::move(bucket);
         }
@@ -373,7 +373,7 @@ class Storage : public IStorage {
       return {Time{}, Error{.code = 422, .message = fmt::format("'{}' parameter can't be empty", param_name)}};
     }
     try {
-      ts = Time{} + std::chrono::microseconds(std::stol(std::string{timestamp}));
+      ts = Time{} + std::chrono::microseconds(std::stoull(std::string{timestamp}));
       return {ts, Error::kOk};
     } catch (...) {
       return {Time{},