+++ unit_tests/reduct/storage/entry_test.cc
@@ -228,12 +228,14 @@ TEST_CASE("storage::Entry should restore itself from folder", "[entry]") {
     std::string_view ts = "1659810772861000";
     const auto meta_path = path / kName / fmt::format("{}.{}", ts, "meta");
     const auto block_path = path / kName / fmt::format("{}.{}", ts, "blk");
-    std::ofstream meta(meta_path);
-    std::ofstream block(block_path);
+    {
+      // Close files (Windows)
+      std::ofstream meta(meta_path);
+      std::ofstream block(block_path);
 
-    auto content = GENERATE("", "garbage");
-    meta << content << std::flush;
-
+      auto content = GENERATE("", "garbage");
+      meta << content << std::flush;
+    }
     entry = IEntry::Build(kName, path, options);
     REQUIRE_FALSE(fs::exists(meta_path));
     REQUIRE_FALSE(fs::exists(block_path));
@@ -402,12 +404,15 @@ TEST_CASE("storage::Entry should wait when read operations finish before removin
 
   REQUIRE(WriteOne(*entry, "blob", kTimestamp) == Error::kOk);
 
-  auto [reader, err] = entry->BeginRead(kTimestamp);
-  REQUIRE(err == Error::kOk);
+  {
+    // reader must be removed
+    auto [reader, err] = entry->BeginRead(kTimestamp);
+    REQUIRE(err == Error::kOk);
 
-  REQUIRE(entry->RemoveOldestBlock() == Error{.code = 500, .message = "Block has active readers"});
-  auto [chunk, read_err] = reader->Read();
-  REQUIRE(read_err == Error::kOk);
+    REQUIRE(entry->RemoveOldestBlock() == Error{.code = 500, .message = "Block has active readers"});
+    auto [chunk, read_err] = reader->Read();
+    REQUIRE(read_err == Error::kOk);
+  }
 
   REQUIRE(entry->RemoveOldestBlock() == Error::kOk);
 }
@@ -416,11 +421,14 @@ TEST_CASE("storage::Entry should wait when write operations finish before removi
   auto entry = IEntry::Build(kName, BuildTmpDirectory(), MakeDefaultOptions());
   REQUIRE(entry);
 
-  auto [writer, err] = entry->BeginWrite(kTimestamp, 5);
-  REQUIRE(err == Error::kOk);
+  {
+    // writer must be removed
+    auto [writer, err] = entry->BeginWrite(kTimestamp, 5);
+    REQUIRE(err == Error::kOk);
 
-  REQUIRE(entry->RemoveOldestBlock() == Error{.code = 500, .message = "Block has active writers"});
-  REQUIRE(writer->Write("12345", true) == Error::kOk);
+    REQUIRE(entry->RemoveOldestBlock() == Error{.code = 500, .message = "Block has active writers"});
+    REQUIRE(writer->Write("12345", true) == Error::kOk);
+  }
 
   REQUIRE(entry->RemoveOldestBlock() == Error::kOk);
 }