+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Build a static executable for AMD64 and upload it to release from CI, [PR-171](https://github.com/reduct-storage/reduct-storage/pull/171)
 - Build on MacOS, [PR-173](https://github.com/reduct-storage/reduct-storage/pull/173)
+- Build on Windows, [PR-174](https://github.com/reduct-storage/reduct-storage/pull/174)
 
 ### Changed: